Refactor modular 1 (#28)
* fix: restore decimal input functionality for quantities

- Re-add quantity_amount_text field to preserve decimal while typing
- Fix issue where '1.' would display as '1' in input fields
- Ensure decimal pad keyboard shows on iOS devices
- Apply fix to both add-item and edit-item screens

* fix: restore food icons display in refactored pantry items

* docs: completely overhaul README with comprehensive project documentation

- Add professional project overview and features
- Include complete tech stack breakdown
- Add visual architecture diagrams
- Improve database schema presentation with tables
- Add step-by-step setup instructions
- Include team collaboration guidelines and git workflow
- Add troubleshooting section for common issues
- Link to new modular architecture documentation
- Add commit conventions and security best practices
